+++ src/pages/Home/index.jsx
@@ -1,22 +1,12 @@
 /* eslint-disable no-console */
 import React, { useState } from 'react';
 import PropTypes from 'prop-types';
-import { Header, Button, Grid, Segment } from 'semantic-ui-react';
+import { Header, Grid } from 'semantic-ui-react';
 import { Redirect } from 'react-router-dom';
 
 import SearchBar from '../../components/SearchBar';
-import ConjugationContainer from '../../components/ConjugationContainer';
-import OptionLabels from '../../components/OptionLabels';
-import ResultSegment from '../../components/ResultSegment';
+import AccentButtons from '../../components/AccentButtons';
 import filterVerbs from '../../helper/filterVerbs';
-
-import info from '../../../globals.json';
-const accentButtons = ['´', '¨', '˜'];
-const toggleAccent = [
-  { á: 'a', é: 'e', í: 'i', ó: 'o', ú: 'u', a: 'á', e: 'é', i: 'í', o: 'ó', u: 'ú', ü: 'ú' },
-  { ü: 'u', u: 'ü', ú: 'ü' },
-  { ñ: 'n', n: 'ñ' },
-];
 
 const Home = () => {
   let [searchValue, setSearchValue] = useState('');
@@ -29,13 +19,6 @@ const Home = () => {
     setRedirect(true);
   };
 
-  const handleAccentClick = (e, accent) => {
-    const cChar = searchValue.slice(-1);
-    const nChar = toggleAccent[accentButtons.indexOf(accent)][cChar];
-    if (nChar) {
-      setSearchValue(prev => prev.slice(0, -1) + nChar);
-    }
-  };
   return redirect ? (
     <Redirect to={`/conjugate/${searchValue}`} />
   ) : (
@@ -57,13 +40,7 @@ const Home = () => {
             autoFocus
           />
           <br />
-          <Button.Group basic size="mini" style={{ margin: '0 auto' }}>
-            {accentButtons.map(elem => (
-              <Button key={elem} onClick={e => handleAccentClick(e, elem)}>
-                {elem}
-              </Button>
-            ))}
-          </Button.Group>
+          <AccentButtons searchValue={searchValue} setSearchValue={setSearchValue} />
         </Grid.Column>
       </Grid.Row>
     </Grid>
